rename index.js to full.js to facilitate merging
